requirements.txt
# Centaur Authors' Proven Working Environment
# Based on: https://github.com/marcelbinz/Llama-3.1-Centaur-70B
# Python 3.10.14 required

# PyTorch with CUDA 12.1 (install separately with index-url)
# torch==2.3.0+cu121
# torchvision (compatible version)
# torchaudio (compatible version)

# Core ML packages - EXACT versions from authors
transformers==4.43.3
trl==0.8.6
peft==0.10.0
accelerate==0.30.1
bitsandbytes==0.43.1

# Flash Attention (proven working with this setup)
xformers==0.0.26.post1

# Data handling and tokenization
datasets==2.19.1
tokenizers==0.19.1
safetensors==0.4.3
dill==0.3.8
multiprocess==0.70.16
pyarrow==16.0.0
pyarrow-hotfix==0.6

# Core scientific computing
numpy==1.26.4
pandas==2.2.2
scipy==1.14.0
scikit-learn==1.5.0

# HuggingFace ecosystem
huggingface-hub==0.26.2

# Essential utilities
tqdm==4.66.4
requests==2.31.0
pyyaml==6.0.1
regex==2024.5.10
sentencepiece==0.2.0
einops==0.8.0

# Development and debugging
ipython==8.26.0
ipdb==0.13.13
jedi==0.19.1

# Visualization
matplotlib==3.9.2
pillow==10.4.0

# Async and networking
aiohttp==3.9.5
aiosignal==1.3.1
frozenlist==1.4.1
multidict==6.0.5
yarl==1.9.4
async-timeout==4.0.3
attrs==23.2.0

# File and data processing
fsspec==2024.3.1
jsonlines==4.0.0
click==8.1.7

# Utilities and parsing
tyro==0.8.4
simple-parsing==0.1.5
inquirerpy==0.3.4
natsort==8.4.0

# Medical/Scientific data (if needed)
nibabel==5.2.1

# Monitoring and logging
wandb==0.17.0
rich==13.7.1

# Version and type checking
typing_extensions==4.11.0
packaging==24.0

# Additional ML utilities
torcheval==0.0.7
schedulefree==1.2.6
ivon-opt==0.1.2

# Security and protocols
certifi==2024.2.2
urllib3==2.2.1
charset-normalizer==3.3.2
idna==3.7

# Git integration
gitpython==3.1.43
gitdb==4.0.11
smmap==5.0.1

# Development tools
decorator==5.1.1
executing==2.0.1
pure-eval==0.2.2
stack-data==0.6.3
six==1.16.0

# Note: Unsloth should be installed separately via git after this
# pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git" 